fix: fixing some elements
