+++ apps/todo/web/src/app/modules/home/pages/index/home-index.module.ts
@@ -1,10 +1,13 @@
 import { NgModule } from '@angular/core';
 import { HomeIndexPage } from './home-index.page';
 import { RouterModule } from '@angular/router';
+import { TodoTitleDirectiveModule } from '../../../todo/directives/todo-title/todo-title.module';
 
 @NgModule({
     declarations: [HomeIndexPage],
     imports: [
+        TodoTitleDirectiveModule,
+        
         RouterModule.forChild([
             {
                 path: '',
