+++ apps/todo/web/src/app/modules/home/pages/index/home-index.page.html
@@ -1,3 +1,7 @@
 <div class="container px-4 md:px-6 py-10">
-    <h1 class="text-2xl font-bold">Todo App</h1>
+    <h1 class="text-2xl font-bold"
+        ubudTodoTitle
+    >
+        Todo App
+    </h1>
 </div>
