+++ apps/todo/web/src/app/modules/todo/containers/todo-list/todo-list.container.html
@@ -1,6 +1,11 @@
 <div class="space-y-4">
     <div>
-        <ubud-element-todo-item-card-web [title]="'Task 1'"></ubud-element-todo-item-card-web>
+        <ubud-element-todo-item-card-web>
+            <ng-container header>
+                <span class="font-bold">Task 1</span>
+            </ng-container>
+            <ng-container content>Lorem ipsum dolor sit amet</ng-container>
+        </ubud-element-todo-item-card-web>
     </div>
     <div>
         <ubud-element-todo-item-card-web [title]="'Task 2'"></ubud-element-todo-item-card-web>
