+++ apps/todo/web/src/app/modules/todo/directives/todo-title/todo-title.module.ts
@@ -1,0 +1,13 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { TodoTitleDirective } from './todo-title.directive';
+
+@NgModule({
+    declarations: [TodoTitleDirective],
+    imports: [
+        CommonModule,
+    ],
+    exports: [TodoTitleDirective],
+})
+export class TodoTitleDirectiveModule {
+}
