+++ apps/todo/web/src/app/modules/todo/pages/index/todo-index.module.ts
@@ -1,12 +1,14 @@
 import { NgModule } from '@angular/core';
-import { TodoIndexPage, TodoIndexTitleDirective } from './todo-index.page';
+import { TodoIndexPage } from './todo-index.page';
 import { RouterModule } from '@angular/router';
 import { TodoListContainerModule } from '../../containers/todo-list/todo-list.module';
+import { TodoTitleDirectiveModule } from '../../directives/todo-title/todo-title.module';
 
 @NgModule({
-    declarations: [TodoIndexPage, TodoIndexTitleDirective],
+    declarations: [TodoIndexPage],
     imports: [
         TodoListContainerModule,
+        TodoTitleDirectiveModule,
         
         RouterModule.forChild([
             {
