+++ apps/todo/web/src/app/modules/todo/pages/index/todo-index.page.html
@@ -1,5 +1,9 @@
 <div class="container px-4 md:px-6 py-10">
-    <h1 class="text-2xl font-bold" ubudTodoIndexTitle>Todo List Page</h1>
+    <h1 class="text-2xl font-bold"
+        ubudTodoTitle
+    >
+        Todo List Page
+    </h1>
     
     <div class="mt-4">
         <ubud-todo-list-container></ubud-todo-list-container>
