+++ apps/todo/web/src/app/modules/todo/pages/index/todo-index.page.ts
@@ -1,25 +1,4 @@
-import { ChangeDetectionStrategy, Component, Directive, ElementRef, HostListener } from '@angular/core';
-
-@Directive({
-    selector: '[ubudTodoIndexTitle]',
-})
-export class TodoIndexTitleDirective {
-    constructor(private el: ElementRef) {
-        // this.el.nativeElement.classList.add('text-primary');
-    }
-
-    @HostListener('mouseenter') onMouseEnter() {
-        this.highlight('yellow');
-    }
-
-    @HostListener('mouseleave') onMouseLeave() {
-        this.highlight('');
-    }
-
-    private highlight(color: string) {
-        this.el.nativeElement.style.backgroundColor = color;
-    }
-}
+import { ChangeDetectionStrategy, Component } from '@angular/core';
 
 @Component({
     selector: 'ubud-todo-index-page',
