+++ apps/todo/web/tailwind.config.js
@@ -3,7 +3,8 @@ const { join } = require('path');
 module.exports = {
     content: [
         join(__dirname, 'src/**/!(*.stories|*.spec).{ts,html}'),
-        join(__dirname, '../../../libs/element/common/**/!(*.stories|*.spec).{ts,html}')
+        join(__dirname, '../../../libs/element/common/**/!(*.stories|*.spec).{ts,html}'),
+        join(__dirname, '../../../libs/element/todo/**/!(*.stories|*.spec).{ts,html}')
     ],
     theme: {
         screens: {
