+++ libs/element/todo/src/lib/modules/components/cards/todo-item/web/todo-item.card.html
@@ -1,3 +1,16 @@
 <div class="border rounded-xl border-grey-200 p-4">
-    {{ title }}
+    <div>
+        <ng-container *ngIf="title else noTitleTpl">
+            {{ title }}
+        </ng-container>
+        <ng-template #noTitleTpl>
+            <div class="header">
+                <ng-content select="[header]"></ng-content>
+            </div>
+
+            <div class="content mt-4">
+                <ng-content select="[content]"></ng-content>
+            </div>
+        </ng-template>
+    </div>
 </div>
